backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.51.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
werkzeug==3.0.1
Pillow==10.1.0
requests==2.31.0
httpx==0.27.0
pyairtable==2.3.3
